refactor(limits): single canonical automation-limits module
